Change the way ov-player directive's attributes were handled and replace ov-fullscreen
All ov-player attributes were handled using different names, use the same names for better understanding.
Also replace the ov-fullscreen attribute by ov-full-viewport which better suits the behaviour. There is now a distinction made between fullscreen and full viewport. Fullscreen use the whole screen if fullscreen API is available. If not a full viewport is performed. While it's not possible to display player in fullscreen (except with the fullscreen button), it is possible to display it in full viewport.