Remove conversion request from the PlayerDirective
When the video points of interest's were needing a
unit conversion from percents to milliseconds. The
player was sending a request to publish to perform
this conversion. However this behaviour creates a
dependency from the player to publish.

To avoid this dependency, the player now triggers
a "needPoiConversion" event and the points of in-
terests won't be parsed if the video needs to be
converted.