Add templates to the distributed files
To integrate the OpenVeo Player it was necessary
to define the global ovPlayerDirectory variable
indicating the URI to the OpenVeo Player sources.
It was used to request the OpenVeo Player
AngularJS templates. AngularJS templates are now
directly embedded into the final JavaScript file
using AngularJS templates cache mechanism. Thus
ovPlayerDirectory is not needed anymore.