Force HTML player to only load metadata of a video
When setting the source of the HTML player
videojs was buffering the video while it is
possible to get only metadata of the video. This
behaviour could lead the HTML page to slow down.
As the OpenVeo Player is designed to be embedded,
an HTML page could hold several OpenVeo Player. To
avoid speed reduction, the HTML player now asks
the browser to only load metadata of the video.