Add output attributes to opl-slider and CSS vars
Output attributes "opl-on-focus", "opl-on-over",
"opl-on-out" and "opl-on-move" have been added.
"opl-on-focus" is called when slider component
receives the focus.
"opl-on-over" is called when a pointer is over the
slider component.
"opl-on-out" is called when a pointer goes out of
the slider component.
"opl-on-move" is called when a pointer moves over
the slider component with information about the
pointer position and the slider value for this
position.

Note that properties "value", "label", "valueText"
and "valueTextParameters" are no longer exposed on
the slider controller.

Finally the slider component now uses CSS
variables, thus colors can be overriden using
--opl-slider-primary-color,
--opl-slider-primary-color-0 and
--opl-slider-accent-color.