Add components to present points of interest
opl-tiles component has been added to present a
list of points of interest as an horizontal list
of tiles. Each point of interest can by of type
"text" or "image".
opl-tiles component is composed of a list of
opl-tile components. It has the size of its parent
element in the DOM tree and scrolls horizontally.
A "text" tile displays a title and a time plus a
"more info" button.
An "image" tile displays an image and a time plus
a "more info" button.
Hitting the "more info" button enlarges the tile
to display more information about it. If tile is
of type "text" it displays the full title, a
scrollable description and an attachment file. If
tile is of type "image" it displays the enlarged
image.