Fix flickering controls
Player controls were flickering when moving the
mouse over the controls elements. It was due to
the mouseout handler which was hiding controls
when a mouseout event was dispatched for all media
sub elements. It now hides controls only if mouse
is moving out the media wrapper.

It also automatically hides controls after 5
seconds of mouse inactivity.