Fix size issues when opening points of interest
Opening a point of interest of type "text" wasn't
working well: the height of the description
placeholder was wrong.

It was due to several caveats due to browser
painting time, too much resets on scroller, tiles
and tile components and the use of the bounding
rectangle returned by getBoundingClientRect which
ignores margins.

It now makes sure that only one reset is runnng at
a time for a component and that clientHeight and
clientWidth properties are used instead of
getBoundingClientRect when necessary.