dashjs: fix JavaScript errors due to upgrade
Since the upgrade of dashjs, two JavaScript errors
appeared. The first one due to changes in the API
thus a JavaScript error was thrown when trying to
deactivate dashjs logs.
The second JavaScript error was thrown when trying
to switch from one source to another, videojs was
asking to set the new source URL then to load it.
However it seems that videojs already load the
source when setting the new source, thus dashjs
was receiving confusing information which didn't
break until version 3.0.3 (it works with version
3.0.2). Consequently videojs is now asked to only
set the new source.