player: fix empty points of interest bar
Points of interest bar could become empty when
changing opl-tags / opl-indexes / opl-chapters
attributes. If tags tab is selected changing the
opl-tags to false led to an empty points of
interest bar even if there were indexes and
chapters. If selected points of interest is masked
it now selects the next available tab.